Move CosmosDB container to a new package - CosmosDB container is moved to a new package in the Azure module - Leave deprecated implementation at old location - Update documentation
