+++ docs/modules/azure.md
@@ -18,31 +18,31 @@ CosmosDBEmulatorContainer | [mcr.microsoft.com/cosmosdb/linux/azure-cosmos-emula
 Start Azure CosmosDB Emulator during a test:
 
 <!--codeinclude-->
-[Starting a Azure CosmosDB Emulator container](../../modules/azure/src/test/java/org/testcontainers/containers/CosmosDBEmulatorContainerTest.java) inside_block:emulatorContainer
+[Starting a Azure CosmosDB Emulator container](../../modules/azure/src/test/java/org/testcontainers/azure/CosmosDBEmulatorContainerTest.java) inside_block:emulatorContainer
 <!--/codeinclude-->
 
 Prepare KeyStore to use for SSL.
 
 <!--codeinclude-->
-[Building KeyStore from certificate within container](../../modules/azure/src/test/java/org/testcontainers/containers/CosmosDBEmulatorContainerTest.java) inside_block:buildAndSaveNewKeyStore
+[Building KeyStore from certificate within container](../../modules/azure/src/test/java/org/testcontainers/azure/CosmosDBEmulatorContainerTest.java) inside_block:buildAndSaveNewKeyStore
 <!--/codeinclude-->
 
 Set system trust-store parameters to use already built KeyStore:
 
 <!--codeinclude-->
-[Set system trust-store parameters](../../modules/azure/src/test/java/org/testcontainers/containers/CosmosDBEmulatorContainerTest.java) inside_block:setSystemTrustStoreParameters
+[Set system trust-store parameters](../../modules/azure/src/test/java/org/testcontainers/azure/CosmosDBEmulatorContainerTest.java) inside_block:setSystemTrustStoreParameters
 <!--/codeinclude-->
 
 Build Azure CosmosDB client:
 
 <!--codeinclude-->
-[Build Azure CosmosDB client](../../modules/azure/src/test/java/org/testcontainers/containers/CosmosDBEmulatorContainerTest.java) inside_block:buildClient
+[Build Azure CosmosDB client](../../modules/azure/src/test/java/org/testcontainers/azure/CosmosDBEmulatorContainerTest.java) inside_block:buildClient
 <!--/codeinclude-->
 
 Test against the Emulator:
 
 <!--codeinclude-->
-[Testing against Azure CosmosDB Emulator container](../../modules/azure/src/test/java/org/testcontainers/containers/CosmosDBEmulatorContainerTest.java) inside_block:testWithClientAgainstEmulatorContainer
+[Testing against Azure CosmosDB Emulator container](../../modules/azure/src/test/java/org/testcontainers/azure/CosmosDBEmulatorContainerTest.java) inside_block:testWithClientAgainstEmulatorContainer
 <!--/codeinclude-->
 
 ## Adding this module to your project dependencies
